+++ lib/src/delegate/future.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library async.delegate.future;
-
 import 'dart:async';
 
 /// A wrapper that forwards calls to a [Future].