+++ lib/src/delegate/sink.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library async.delegate.sink;
-
 /// Simple delegating wrapper around a [Sink].
 ///
 /// Subclasses can override individual methods, or use this to expose only the