+++ lib/src/delegate/stream_subscription.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library async.delegate.stream_subscription;
-
 import 'dart:async';
 
 /// Simple delegating wrapper around a [StreamSubscription].