+++ lib/src/null_stream_sink.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library async.null_stream_sink;
-
 import 'dart:async';
 
 /// A [StreamSink] that discards all events.