+++ lib/src/restartable_timer.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library async.restartable_timer;
-
 import 'dart:async';
 
 /// A non-periodic timer that can be restarted any number of times.