+++ lib/src/stream_sink_completer.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library async.stream_sink_completer;
-
 import 'dart:async';
 
 import 'null_stream_sink.dart';