+++ test/utils.dart
@@ -3,8 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 /// Helper utilities for testing.
-library async.test.util;
-
 import "dart:async";
 
 import "package:async/async.dart";